DO NOT MERGE - tc-pulse creds work in progress
Work is not complete as I have run out of time. The namespace contact information Lines 89-98 must
be parameterized. I ran into issues matching type signatures of the pulse.SendEmailRequest struct
and was unable to complete the work before the end of my term.
This code is currently work in progress and has not been tested.